backend/requirements.txt
click
Flask
Flask-JWT-Extended
itsdangerous
Jinja2
MarkupSafe
flask-cors
PyJWT
python-dotenv
Werkzeug
flask-sqlalchemy
flask-migrate
flask-bcrypt
gunicorn
psycopg2-binary